Able to connect to the CloudyWeather API and log current conditions.
A Suspender utility function has been created that throws a promise if
the data is not available. This allows the component to suspend while
the data is being fetched. We can use this with the Suspense API and
display a Fallback component until the data is ready. This method
eliminates the double fetch on load using the useEffect hook. While
there are remedies for this, the Suspender utility function is a cleaner
way to handle data fetching.