We can get the user's location by GPS or IP address. We default to GPS
data because it is more accurate. If the user denies location services
with their browser, we fall back to IP address. IF there is an error
fetching the geo-location data this way we alert the user to enable
location services be we will default them to 0,0 or possibly random
coordinates in the future.