Working on the CurrentConditions component and its sub-components
A LocationHeader component has been added to manage the location's name
and the time. The time is updated every 5 seconds to reduce overall
events being fired. We don't need it to be overly accurate. Helpers for
the component have also been created to help with formatting or
gathering data to display.

A useIsMounted Hook has been created to keep the repeated logic dry. As
many components will need to ensure they are mounted before updating
state.

The suspender has been updated to be reuseable and now accepts a T
type. This will allow us to use it for any type of data we want to
fetch.

The remaining data except for custom icons we already have so the
next round of work should be relatively simple to complete.